Add `Pipe` for function composition
| Pipe is a mechanism or an operator that allows you to compose
| functions together in a sequence, passing the output of one function
| as the input to the next function. Function composition enables you to
| build complex computations by chaining together simpler functions,
| making your code more concise and readable.

Add Pipe as a separate package `pipe` with a README to explain briefly
the reason for Pipe and it's usage.

Add section `Concepts` to main README to collect all concepts in root
location.

To create a composition call PipeN, the N as the amount of given
function arguments. The initial value is passed when calling the
composed function. The type for the initial value is inferred from the
first function argument.