Add `State` monad
| The State monad is a programming construct often used in functional
| programming languages to manage stateful computations in a pure and
| immutable manner. It provides a way to encapsulate stateful operations
| while maintaining referential transparency. State monad represents
| computations that take an initial state as an input and returns the
| result with the updated state as a tuple `(Result, State)`.

Add State monad as a separate package `state` with a README that
explains briefly the State monad rationale and usage.

To create stateful computations we need to define the initial state `a` first
to run computations on. Initial state `s` can be anything, but it must be
wrapped into a State monad.

After wrapping the state into the State monad we can run stateful computations
on it for exampel by using `state.Modify` and `state.Put`.

Before running stateful operations we need to Run the state to access the
underlying state operation. `pipe.Pipe2(state.Run[int], state.Map[int](increment))(initialState)`